beginner-html-site-styled-gh-pages
To create a basic HTML website that is styled and hosted on GitHub Pages, you will need to do the following:

Create a new repository on GitHub.
In the repository, create an index.html file. This will be the main page of your website.
Add your HTML code to the index.html file. You can use a basic template like this to get started:
Copy code
<!DOCTYPE html>
<html>
<head>
  <title>My Website</title>
</head>
<body>
  <h1>Welcome to my website!</h1>
  <p>This is my first website.</p>
</body>
</html>
Create a style.css file in the repository to hold your CSS code.
Add a link to the style.css file in the head of your index.html file:
Copy code
<link rel="stylesheet" type="text/css" href="style.css">
Add your CSS code to the style.css file to style your website.

Commit your changes and push them to your repository.

Go to the Settings tab of your repository, and under the GitHub Pages section, select the master branch as the source for your website.

Your website should now be live and available at the URL provided in the GitHub Pages section.